worker/requirements.txt
supabase>=2.4.0

yt-dlp>=2024.1.7
ffmpeg-python==0.2.0
python-dotenv==1.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
